requirements.txt
allure-pytest
allure-python-commons
apipkg
appdirs
astroid
async-generator
atomicwrites
attrs
autopep8
Babel
backports.entry-points-selectable
beautifulsoup4
bidict
blinker
bs4
cachelib
cachetools
certifi
cffi
chardet
charset-normalizer
click
colorama
configparser
crayons
cycler
decorator
deepdiff
distlib
dnspython
docxcompose
docxtpl
exceptiongroup
execnet
Faker
filelock
Flask
Flask-Ext
Flask-Mail
flask-marshmallow
Flask-MySQL
Flask-MySQLdb
flask-session
Flask-SocketIO
flask-sqlalchemy
flask-swagger-ui
fonttools
google
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
greenlet
gunicorn
h11
helium
httplib2
idna
importlib-metadata
iniconfig
isort
itsdangerous
Jinja2
jmespath
jsonpath
jsonpath-ng
jsonschema
kiwisolver
lazy-object-proxy
lxml
MarkupSafe
marshmallow
matplotlib
mccabe
MouseInfo
mysql
mysql-connector-python
mysqlclient
numpy
oauthlib
ordered-set
packaging
passlib
Pillow
pluggy
ply
protobuf
pyasn1
pyasn1-modules
pycodestyle
pycparser
PyJWT
PyMySQL
pyparsing
pyperclip
pyrsistent
pytest
python-dateutil
python-docx
python-dotenv
python-engineio
python-socketio
pytz
requests
requests-oauthlib
rsa
selenium
simple-websocket
six
soupsieve
SQLAlchemy
text-unidecode
tk
toml
tomli
tqdm
typed-ast
typing-extensions
uritemplate
urllib3
webdriver-manager
Werkzeug
wget
wrapt
wsproto
WTForms
zipp
mkdocs
faiss-cpu
python-dotenv
lxml
greenlet
typed-ast
openai==0.28.0
